+++ examples/Server/Modules/Orders/Application/HubCommandHandlers/UnsubscribeFromSymbolHubCommandHandler.cs
@@ -1,17 +1,14 @@
 using ManagedDotNet.SignalR.Topics.Examples.Server.Modules.Orders.Infrastructure;
 using ManagedDotNet.SignalR.Topics.Examples.Server.Modules.Orders.Models;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.Logging;
 using ManagedDotNet.SignalR.Topics.Abstractions;
 using ManagedDotNet.SignalR.Topics.Core;
 using ManagedDotNet.SignalR.Topics.Examples.Server.Modules.Orders;
-using static ManagedDotNet.SignalR.Topics.Examples.Shared.Services.AuthService;
 
 namespace ManagedDotNet.SignalR.Topics.Examples.Server.Modules.Orders.Application.HubCommandHandlers;
 
 
-[Authorize(Roles = $"{Roles.User},{Roles.Administrator}")]
 public class UnsubscribeFromSymbolHubCommandHandler : IHubCommandHandler<UnsubscribeFromSymbolCommand>
 {
     private readonly ITopicHubContext<OrderBookHub> _hubContext;